stream_processor/requirements.txt
kafka-python==2.0.2
redis==4.5.5
numpy==1.26.4
pandas==2.1.4
joblib==1.3.2
